+++ include/simgrid/s4u/Host.hpp
@@ -85,7 +85,7 @@ public:
 
   kernel::routing::NetPoint* get_netpoint() const { return pimpl_netpoint_; }
 
-  int get_actor_count() const;
+  size_t get_actor_count() const;
   std::vector<ActorPtr> get_all_actors() const;
 
   /** Turns that host on if it was previously off
